% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{.onUnload}
\alias{.onUnload}
\title{Unload dll when package is unloaded}
\usage{
.onUnload(libpath)
}
\value{
No return value, called for side effects.
}
\description{
Unload dll when package is unloaded
}
\keyword{internal}
