% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivHazardLoop.R
\name{multivHazardLoop}
\alias{multivHazardLoop}
\title{Loop for multivariate estimation under hazard ratio ordering}
\usage{
multivHazardLoop(X, thresholds, nThr, weights, cpY, pars)
}
\description{
Loop for multivariate estimation under hazard ratio ordering
}
\keyword{internal}
