% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{plain_survival_isotonic_distributional_regression}
\alias{plain_survival_isotonic_distributional_regression}
\title{Compute the plain survival IDR for totally ordered co-variates under the hazard rate order
assumption for censored data. Note that the hazard rate order is needed for the method to be
consistent, but that the returned solution likely doesn't satisfy this assumption.}
\usage{
plain_survival_isotonic_distributional_regression(X, y, y_observed, weights)
}
\arguments{
\item{X}{Double vector of scalar covariates.}

\item{y}{Double vector of response values.}

\item{y_observed}{Integer vector: 1 for observed, 0 for censored.}

\item{weights}{Double vector of non-negative weights. All vectors must have equal length.}
}
\value{
Numeric matrix. The (i, j) entry gives the fitted CDF at (unique) response j for
  (unique) covariate i.
}
\description{
Computes the plain survival isotonic distributional regression (plain survival IDR) under the
hazard rate order assumption for totally ordered covariates when some responses are right
censored. Returns the fitted cumulative distribution evaluated at each response threshold for
each unique covariate.
}
\details{
This method is provided for completeness and to make the S-IDR publication easier to reproduce,
but is not recommended for use by practitioners due to prohibitive computational costs and
potential inconsistency in case the distributions are not hazard rate ordered.
}
\examples{
plain_survival_isotonic_distributional_regression(as.double(1:4), c(2, 1, 4, 3),
  as.integer(c(0, 1, 0, 1)), c(1, 2, 1, 1))
}
