% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{prepareData}
\alias{prepareData}
\title{Prepare data for IDR modeling with given orders}
\usage{
prepareData(X, groups, orders)
}
\arguments{
\item{X}{\code{data.frame} of covariates.}

\item{groups}{vector of groups.}

\item{orders}{named vector of orders for groups.}
}
\value{
Transformed data \code{X} with covariates modified according to the given
order.
}
\description{
Prepare data for IDR modeling with given orders
}
\keyword{internal}
