% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialOrders.R
\name{splitArr}
\alias{splitArr}
\title{Split array/matrix by margins (for R before and after 3.6)}
\usage{
splitArr(x, MARGIN, asplitAvail = TRUE)
}
\arguments{
\item{x}{an array, including a matrix.}

\item{MARGIN}{a vector giving the margins to split.}

\item{asplitAvail}{choose \code{\link[base]{asplit}} (available in 
R >= 3.6), otherwise \code{\link[base]{split}} (less efficient).}
}
\value{
A list array splitted along the given \code{MARGIN}.
}
\description{
Wrapper for \code{\link[base]{asplit}} and
\code{\link[base]{split}}.
}
\keyword{internal}
