% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{survival_isotonic_distributional_regression_threshold}
\alias{survival_isotonic_distributional_regression_threshold}
\title{Compute an isotonic regression of a probability given by Kaplan-Meier estimators (= one
threshold of S-IDR).}
\usage{
survival_isotonic_distributional_regression_threshold(
  threshold,
  X,
  y,
  y_observed,
  weights,
  decreasing = FALSE
)
}
\arguments{
\item{threshold}{Double of the response value at which to compute the IDR solution.}

\item{X}{Double vector of totally ordered covariates.}

\item{y}{Double vector of response values.}

\item{y_observed}{Integer vector: 1 for observed, 0 for censored.}

\item{weights}{Double vector of non-negative weights. All vectors must have equal length.}

\item{decreasing}{Bool indicating direction (decreasing is a CIDR threshold).}
}
\value{
Numeric vector. The i'th entry gives the fitted CDF at covariate i.
}
\description{
This method is provided for completeness and to make the S-IDR publication easier to reproduce.
}
\examples{
survival_isotonic_distributional_regression_threshold(3.5, as.double(1:4), c(2, 1, 4, 3),
  as.integer(c(1, 0, 0, 0)), rep(1.0, 4))
}
