% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialOrders.R
\name{trReduc}
\alias{trReduc}
\title{Transitive Reduction of path matrix}
\usage{
trReduc(paths, n)
}
\arguments{
\item{paths}{a two column integer matrix containing all pairs of nodes
which are linked by a path of edges.}

\item{n}{the number of nodes.}
}
\value{
A two column matrix giving all pairs of edges \code{(i,j)} such that there
exists a directed edge from node \code{i} to node \code{j}.
}
\description{
Computes the transitive reduction of the path matrix of a directed acyclic
graph.
}
\details{
For each \code{k}, all indirect paths going through node \code{k} are
removed (if there are any).
}
\keyword{internal}
