% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{warn_once}
\alias{warn_once}
\title{Helper to warn once per session
Warn only once per session}
\usage{
warn_once(key, msg, call. = FALSE, immediate. = FALSE)
}
\arguments{
\item{key}{How to identify the warning.}

\item{msg}{Message text.}

\item{call.}{Logical; if \code{TRUE}, the call is included in the warning
message.}

\item{immediate.}{Logical; if \code{TRUE}, the warning is issued
immediately rather than being deferred.}
}
\description{
Filters warnings, letting through only the first of each.
}
